requirements.txt
docker-py==1.8.1
ipython==5.0.0